Move the tests of the deprecated --pre-commit argument to a dedicated file
